+++ codegen/src/generators/ResourcesGenerator/httpVerbPathOperations.ts
@@ -1,0 +1,13 @@
+/**
+ * A list of all http verbs that are supported by the OpenApi path.
+ */
+export const httpVerbPathOperations = [
+    'get' as 'get',
+    'put' as 'put',
+    'post' as 'post',
+    'delete' as 'delete',
+    'options' as 'options',
+    'head' as 'head',
+    'patch' as 'patch',
+    'trace' as 'trace',
+]