+++ example/src/SlushyFactory.ts
@@ -1,12 +1,16 @@
 import { Slushy, SlushyConfig, SlushyPlugins } from '@slushy/server'
-import { ResourceConfig } from './generated/resources'
 import { PetsResourceImpl } from './PetsResourceImpl'
+import { ResourcesConfiguration } from './generated/ResourcesConfiguration'
+import { Context } from './Context'
 
 export class SlushyFactory {
-    public static async create(config: Partial<SlushyConfig> & Partial<SlushyPlugins> = {}) {
-        const slushy = await Slushy.create({
+    public static async create(
+        config: Partial<SlushyConfig<Context>> & Partial<SlushyPlugins> = {}
+    ): Promise<Slushy<Context>> {
+        const slushy = await Slushy.create<Context>({
+            contextFactory: ctx => ({ ...ctx, context: {} }),
             ...config,
-            resourceConfiguration: new ResourceConfig({
+            resourceConfiguration: new ResourcesConfiguration({
                 PetsResource: new PetsResourceImpl(),
             }),
         })