+++ server/src/Slushy.ts
@@ -1,16 +1,17 @@
-import { fs } from 'mz'
+// import { fs } from 'mz'
 import { SlushyRouter } from './SlushyRouter'
 import { SlushyApplication, SlushyApplicationFactory } from './ServerImpl'
 import { SlushyProps } from './SlushyProps'
 import { SlushyConfig } from './SlushyConfig'
 import { SlushyPlugins } from './SlushyPlugins'
 import { DefaultLoggerFactory } from './LoggerFactory'
+import { OpenAPIV3 } from 'openapi-types'
 
-export class Slushy {
+export class Slushy<TContext> {
     public constructor(
-        public readonly props: Readonly<SlushyProps>,
+        public readonly props: Readonly<SlushyProps<TContext>>,
         public readonly app: SlushyApplication = SlushyApplicationFactory.create(),
-        public readonly router: SlushyRouter = new SlushyRouter(props, app)
+        public readonly router: SlushyRouter<TContext> = new SlushyRouter(props, app)
     ) {}
 
     public async start(port: number) {
@@ -21,8 +22,8 @@ export class Slushy {
         await this.props.resourceConfiguration.configure(this.router)
     }
 
-    public static async create(config: SlushyConfig & Partial<SlushyPlugins>) {
-        const openApi = JSON.parse((await fs.readFile(config.resourceConfiguration.getOpenApiPath())).toString())
+    public static async create<TContext>(config: SlushyConfig<TContext> & Partial<SlushyPlugins>) {
+        const openApi = JSON.parse(config.resourceConfiguration.getOpenApiSchema()) as OpenAPIV3.Document
         const slushy = new Slushy({
             openApi,
             loggerFactory: new DefaultLoggerFactory(),