+++ server/src/SlushyProps.ts
@@ -2,6 +2,6 @@ import { OpenAPIV3 } from 'openapi-types'
 import { SlushyConfig } from './SlushyConfig'
 import { SlushyPlugins } from './SlushyPlugins'
 
-export interface SlushyProps extends SlushyConfig, SlushyPlugins {
+export interface SlushyProps<TContext> extends SlushyConfig<TContext>, SlushyPlugins {
     openApi: OpenAPIV3.Document
 }