+++ server/src/SlushyResourceConfiguration.ts
@@ -1,6 +1,6 @@
 import { SlushyRouter } from './SlushyRouter'
 
-export interface SlushyResourceConfiguration {
-    configure(router: SlushyRouter): Promise<void>
-    getOpenApiPath(): string
+export interface SlushyResourceConfiguration<TContext> {
+    configure(router: SlushyRouter<TContext>): Promise<void>
+    getOpenApiSchema(): string
 }