+++ server/src/middleware/BodyParser.ts
@@ -3,7 +3,7 @@ import bodyParser from 'body-parser'
 import { SlushyProps } from '../SlushyProps'
 
 export class BodyParser implements MiddlewareFactory {
-    create(_props: SlushyProps) {
+    create(_props: SlushyProps<any>) {
         return [bodyParser.json()]
     }
 }