+++ server/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/aimed/slushy/compare/v0.8.3...v0.9.0) (2019-06-23)
+
+
+### Features
+
+* add context ([#56](https://github.com/aimed/slushy/issues/56)) ([8aad61e](https://github.com/aimed/slushy/commit/8aad61e))
+
+
+
+
+
+## [0.8.2](https://github.com/aimed/slushy/compare/v0.8.1...v0.8.2) (2019-06-22)
+
+**Note:** Version bump only for package @slushy/server
+
+
+
+
+
+## [0.8.1](https://github.com/aimed/slushy/compare/v0.8.0...v0.8.1) (2019-06-22)
+
+**Note:** Version bump only for package @slushy/server
+
+
+
+
+
+# [0.8.0](https://github.com/aimed/slushy/compare/v0.7.1...v0.8.0) (2019-06-06)
+
+**Note:** Version bump only for package @slushy/server
+
+
+
+
+
 # [0.7.0](https://github.com/aimed/slushy/compare/v0.6.0...v0.7.0) (2019-06-06)
 
 **Note:** Version bump only for package @slushy/server